test edit funtion
